requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
python-dotenv==1.0.0
python-docx==0.8.11
PyPDF2==3.0.1
spacy==3.7.2
scikit-learn==1.3.0
pandas==2.0.3
openpyxl==3.1.2
Werkzeug==2.3.7
email-validator==2.0.0.post2
python-magic-bin==0.4.14
nltk==3.8.1
textstat==0.7.3
